py/requirements.txt
altgraph==0.17.4
packaging==24.1
pefile==2024.8.26
pygame==2.6.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
setuptools==74.1.2
